+++ test/build.test.js
@@ -50,3 +50,122 @@ tape('Compile unit tests', function(t) {
   t.end();
 
 });
+
+tape('Compile SNS rule', function(t) {
+  var snsConfig = {
+    name: 'sns',
+    sourcePath: 'rules/sns.js',
+    parameters: {
+      'token': {
+        'Type': 'String',
+        'Description': 'token'
+      }
+    },
+    snsRule: {}
+  };
+
+  var snsBuilt = lambdaCfn.compile([lambdaCfn.build(snsConfig)], {});
+  var snsFixture = JSON.parse(fs.readFileSync(path.join(__dirname, './fixtures/sns.template'), "utf8"));
+  t.deepEqual(snsBuilt,snsFixture, 'SNS rule build is equal to fixture');
+
+  t.end();
+});
+
+
+tape('Compile Event rule', function(t) {
+  var eventConfig = {
+    name: 'eventRule',
+    sourcePath: 'rules/eventRule.js',
+    parameters: {
+      'token': {
+        'Type': 'String',
+        'Description': 'token'
+      }
+    },
+    eventRule: {
+      eventPattern:{
+        'detail-type': [
+          'AWS API Call via CloudTrail'
+        ],
+        detail: {
+          eventSource: [
+            'iam.amazonaws.com'
+          ],
+          eventName: [
+            'CreatePolicy',
+            'CreatePolicyVersion',
+            'PutGroupPolicy',
+            'PutRolePolicy',
+            'PutUserPolicy'
+          ]
+        }
+      }
+    }
+  };
+
+  var eventBuilt = lambdaCfn.compile([lambdaCfn.build(eventConfig)], {});
+  var eventFixture = JSON.parse(fs.readFileSync(path.join(__dirname, './fixtures/event.template'), "utf8"));
+  t.deepEqual(eventBuilt,eventFixture, 'Event rule build is equal to fixture');
+
+  t.end();
+});
+
+tape('Compile Scheduled rule', function(t) {
+  var scheduledConfig = {
+    name: 'scheduledRule',
+    sourcePath: 'rules/scheduledRule.js',
+    parameters: {
+      'token': {
+        'Type': 'String',
+        'Description': 'token'
+      }
+    },
+    scheduledRule: 'rate(5 minutes)'
+  };
+
+  var scheduledBuilt = lambdaCfn.compile([lambdaCfn.build(scheduledConfig)], {});
+  var scheduledFixture = JSON.parse(fs.readFileSync(path.join(__dirname, './fixtures/scheduled.template'), "utf8"));
+  t.deepEqual(scheduledBuilt,scheduledFixture, 'Scheduled rule build is equal to fixture');
+
+  t.end();
+});
+
+
+tape('Compile Hybrid Scheduled and Hybrid based rule', function(t) {
+  var hybridConfig = {
+    name: 'hybridRule',
+    sourcePath: 'rules/hybridRule.js',
+    parameters: {
+      'token': {
+        'Type': 'String',
+        'Description': 'token'
+      }
+    },
+    eventRule: {
+      eventPattern:{
+        'detail-type': [
+          'AWS API Call via CloudTrail'
+        ],
+        detail: {
+          eventSource: [
+            'iam.amazonaws.com'
+          ],
+          eventName: [
+            'CreatePolicy',
+            'CreatePolicyVersion',
+            'PutGroupPolicy',
+            'PutRolePolicy',
+            'PutUserPolicy'
+          ]
+        }
+      }
+    },
+    scheduledRule: 'rate(5 minutes)'
+  };
+
+  var hybridBuilt = lambdaCfn.compile([lambdaCfn.build(hybridConfig)], {});
+  var hybridFixture = JSON.parse(fs.readFileSync(path.join(__dirname, './fixtures/hybrid.template'), "utf8"));
+  t.deepEqual(hybridBuilt,hybridFixture, 'Hybrid rule build is equal to fixture');
+
+  t.end();
+});